+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Removed
 
 ### Fixed
+- Prevent a NPE if the user, group or working directory of a process is unknown.
 
 ## [1.0.0] - 2022-07-02
 ### Added